feature: update Order and OrderItem entities to use BigDecimal for price and add order total calculation

+++ src/main/kotlin/com/projects/orderitem/controller/request/OrderItemRequest.kt
@@ -1,7 +1,9 @@
 package com.projects.orderitem.controller.request
 
+import java.math.BigDecimal
+
 data class OrderItemRequest(
     val productName: String,
     val quantity: Int,
-    val price: Double
+    val price: BigDecimal
 )
